+++ test/fasta/specimens.jl
@@ -1,3 +1,6 @@
+@testset "Specimens" begin
+
+
 @testset "Valid specimens" begin
     # All valid specimens should be read, written, re-read, and the
     # second read should be identical.
@@ -43,3 +46,5 @@ end
         @test !isnothing(open(validate_fasta, path))
     end
 end
+
+end # testset Specimens
