+++ test/fastq/record.jl
@@ -1,3 +1,5 @@
+@testset "Record" begin
+
 # Only using empty records here
 @testset "Basic properties" begin
     # Equality of empty records
@@ -21,32 +23,6 @@
     @test sequence(record) == ""
     @test isempty(quality(record))
 end
-
-TEST_RECORD_STRINGS = [
-    # Standard records
-    "@some_header\r\nAAGG\r\n+\r\njjll",
-    "@prkl_19900 [a b]:211\nkjmn\n+\naabb",
-    "@some_header\nAAGG\n+some_header\njjll\n\n", # same as #1
-
-    # Edge cases:
-    "@\nTAG\n+\n!!!", # empty description
-    "@ ||;;211name \nkakana\n+\naabbcc", # empty some_identifier
-    "@header here\n\n+\n", # empty sequence
-
-    # 
-]
-
-TEST_BAD_RECORD_STRINGS = [
-    "@some\n\nTAG\n+\r\njjj", # extra newline
-    "@abc\nABC\n+\nABCD", # qual too long,
-    "@abc\nABC\n+\nAB", # qual too short,
-    "@A B \nC\n+A B\nA", # second header different
-    "@A\nC\n+AB\nA", # second header too long
-    "@AB\nC\n+A\nA", # second header too short,
-    "@AB\nC\n+AB\n\t", # qual not in range
-    "@AB\nABC\n+\nK V", # qual not in range
-    "@AB\nABC\n+\nK\x7fV", # qual not in range
-]
 
 @testset "Basic construction" begin
     function test_is_equal(a::Record, b::Record)
@@ -293,3 +269,5 @@ end
     @test hash(cp) == hash(records[2])
     @test isequal(cp, records[2])
 end
+
+end # testset Record
